Revert "Allow SerializationTraits to use grpc::ByteBuffer rather than only grpc_byte_buffer"
