+++ include/grpc++/impl/codegen/rpc_service_method.h
@@ -25,10 +25,13 @@
 #include <memory>
 #include <vector>
 
-#include <grpc++/impl/codegen/byte_buffer.h>
 #include <grpc++/impl/codegen/config.h>
 #include <grpc++/impl/codegen/rpc_method.h>
 #include <grpc++/impl/codegen/status.h>
+
+extern "C" {
+struct grpc_byte_buffer;
+}
 
 namespace grpc {
 class ServerContext;
@@ -40,14 +43,11 @@ class MethodHandler {
   virtual ~MethodHandler() {}
   struct HandlerParameter {
     HandlerParameter(Call* c, ServerContext* context, grpc_byte_buffer* req)
-        : call(c), server_context(context) {
-      request.set_buffer(req);
-    }
-    ~HandlerParameter() { request.Release(); }
+        : call(c), server_context(context), request(req) {}
     Call* call;
     ServerContext* server_context;
-    // Handler required to destroy these contents
-    ByteBuffer request;
+    // Handler required to grpc_byte_buffer_destroy this
+    grpc_byte_buffer* request;
   };
   virtual void RunHandler(const HandlerParameter& param) = 0;
 };
