+++ src/core/lib/iomgr/resource_quota.c
@@ -22,7 +22,6 @@
 #include <stdint.h>
 #include <string.h>
 
-#include <grpc/slice_buffer.h>
 #include <grpc/support/alloc.h>
 #include <grpc/support/log.h>
 #include <grpc/support/string_util.h>
