+++ src/cpp/server/health/default_health_check_service.cc
@@ -20,7 +20,6 @@
 #include <mutex>
 
 #include <grpc++/impl/codegen/method_handler_impl.h>
-#include <grpc/slice.h>
 #include <grpc/support/alloc.h>
 #include <grpc/support/log.h>
